Change optional grep-linters to act like required grep-linters.
This makes it easy to use them as templates for required linters.

Normally, grep returns 0 (i.e.: success) if it finds a match, and 1
(i.e.: failure) if it does not find a match. But the debug and TODO
linters are searching for things that should NOT be in the code, i.e.:
the opposite of grep's default behavior.

This change swaps the default behavior, i.e.: returns success when grep
doesn't find anything; and a failure when grep does find something.

How did this work before? They were optional linters, so their output
value is being ignored (both before and after this change). But if you
copied one of these three linters and modified it yourself to become a
required linter, it would have acted the opposite of what you expected.